refactor: 미션 진행에 필요가 없는 hands-on 제거 및 BaseTimeEntity qna 패키지로 이동
