--- src/main/java/favorite/domain/Favorite.java
@@ -1,16 +1,0 @@
-package favorite.domain;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class Favorite {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    protected Favorite() {
-    }
-}
