--- src/main/java/subway/domain/Line.java
@@ -1,47 +1,0 @@
-package subway.domain;
-
-import java.util.ArrayList;
-import java.util.List;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-
-@Entity
-public class Line {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @OneToMany(mappedBy = "line")
-    private List<Station> stations = new ArrayList<>();
-
-    protected Line() {
-    }
-
-    public Line(final String name) {
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public List<Station> getStations() {
-        return stations;
-    }
-
-    public void addStation(final Station station) {
-        this.stations.add(station);
-        station.setLine(this);
-    }
-}
