--- src/main/java/subway/domain/LineRepository.java
@@ -1,7 +1,0 @@
-package subway.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface LineRepository extends JpaRepository<Line, Long> {
-    Line findByName(String name);
-}
