--- src/main/java/subway/domain/Station.java
@@ -1,55 +1,0 @@
-package subway.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-
-@Entity
-public class Station {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @ManyToOne
-    private Line line;
-
-    protected Station() {
-    }
-
-    public Station(final Long id, final String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Station(final String name) {
-        this(null, name);
-    }
-
-    public void changeName(final String name) {
-        this.name = name;
-    }
-
-    public void setLine(final Line line) {
-        this.line = line;
-        line.getStations().add(this);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Line getLine() {
-        return line;
-    }
-}
-
