--- src/main/java/subway/domain/StationRepository.java
@@ -1,7 +1,0 @@
-package subway.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StationRepository extends JpaRepository<Station, Long> {
-    Station findByName(String name);
-}
