--- src/main/resources/subway-data.sql
@@ -1,4 +1,0 @@
-INSERT INTO line (id, name)
-VALUES (1, '3호선');
-INSERT INTO station (id, line_id, name)
-VALUES (1, 1, '교대역')
