--- src/test/java/subway/domain/StationRepositoryTest.java
@@ -1,48 +1,0 @@
-package subway.domain;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-
-@DataJpaTest
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-class StationRepositoryTest {
-    @Autowired
-    private StationRepository stations;
-
-    @Test
-    void save() {
-        final Station expected = new Station("잠실역");
-        final Station actual = stations.save(expected);
-
-        assertThat(actual.getId()).isNotNull();
-        assertThat(actual.getName()).isEqualTo("잠실역");
-    }
-
-    @Test
-    void findByName() {
-        stations.save(new Station("잠실역"));
-        final Station actual = stations.findByName("잠실역");
-        assertThat(actual).isNotNull();
-    }
-
-    @Test
-    void identity() {
-        final Station station1 = stations.save(new Station("잠실역"));
-        final Station station2 = stations.findById(station1.getId()).get();
-
-        assertThat(station1 == station2).isTrue();
-        assertThat(station1).isSameAs(station2);
-    }
-
-    @Test
-    void update() {
-        final Station station1 = stations.save(new Station("잠실역"));
-        station1.changeName("몽촌토성역");
-        final Station station2 = stations.findByName("몽촌토성역");
-        assertThat(station2).isNotNull();
-    }
-}
